requirements.txt
discord.py==2.3.2
fastapi==0.110.0
uvicorn==0.29.0